refactor(scheduler): extract consolidation relation and log rotation handlers (issue #224 task 4)
